fix: 테스트코드 laps 데이터타입 적용 (#116)

+++ module-domain/src/test/java/com/depromeet/service/MemoryServiceTest.java
@@ -121,9 +121,9 @@ class MemoryServiceTest {
     void 회원은_영법을_수정할_수_있다() {
         // given
         UpdateStrokeCommand firstCommand =
-                new UpdateStrokeCommand((Long) null, "자유형", (short) 4, null);
+                new UpdateStrokeCommand((Long) null, "자유형", 4F, null);
         UpdateStrokeCommand secondCommand =
-                new UpdateStrokeCommand((Long) null, "접영", (short) 2, null);
+                new UpdateStrokeCommand((Long) null, "접영", 2F, null);
 
         List<UpdateStrokeCommand> commands = List.of(firstCommand, secondCommand);
