+++ module-infrastructure/persistence-database/src/test/java/com/depromeet/fixture/memory/StrokeFixture.java
@@ -5,7 +5,7 @@ import com.depromeet.memory.domain.Stroke;
 
 public class StrokeFixture {
     public static Stroke mockLapsStroke(Memory memory) {
-        return Stroke.builder().memory(memory).name("자유형").laps((short) 5).build();
+        return Stroke.builder().memory(memory).name("자유형").laps(5F).build();
     }
 
     public static Stroke mockMeterStroke(Memory memory) {
